pyproject.toml
[project]
name = "crypto-data-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]
